refactor(runtime): extract artifact-spill helpers into spill module
resolve_spill_config + spill_or_passthrough move into
tool_runner/spill.rs. These are the shared overflow helpers that
web_fetch and web_search use to spill oversize tool results into the
artifact store (#3347 5/N) instead of burning context. Both are
verbatim moves with visibility upgraded to pub(super); their behavior
on the `0` -> compiled-default fallback in resolve_spill_config is
unchanged.

The stale `// Web tools` section header in mod.rs (which used to sit
above these helpers but no longer aligned with anything after the
upstream web tools moved out earlier) is removed in the same hunk.

mod.rs: 9,219 -> 9,172 LOC (-47). cargo check + cargo clippy --lib
both clean.

Refs #3710
